requirements.txt
pytest==8.3.5
playwright==1.50.0
pytest-playwright==0.7.2
requests==2.32.3
pydantic==2.10.6
faker==33.3.1
psycopg2-binary==2.9.10
allure-pytest==2.13.5
python-dotenv==1.0.1
ruff==0.9.6
mypy==1.19.1
pre-commit==4.5.1

